# scraping: hoist labels variable to save garbage